refactor(remote-config): remove deprecated namespaced API
