+++ packages/app/lib/common/index.ts
@@ -336,28 +336,6 @@ const mapOfDeprecationReplacements: DeprecationMap = {
       nanoseconds: NO_REPLACEMENT,
     },
   },
-  remoteConfig: {
-    default: {
-      activate: 'activate()',
-      ensureInitialized: 'ensureInitialized()',
-      fetchAndActivate: 'fetchAndActivate()',
-      getAll: 'getAll()',
-      getBoolean: 'getBoolean()',
-      getNumber: 'getNumber()',
-      getString: 'getString()',
-      getValue: 'getValue()',
-      reset: 'reset()',
-      setConfigSettings: 'setConfigSettings()',
-      fetch: 'fetch()',
-      setDefaults: 'setDefaults()',
-      setDefaultsFromResource: 'setDefaultsFromResource()',
-      onConfigUpdated: 'onConfigUpdated()',
-    },
-    statics: {
-      LastFetchStatus: 'LastFetchStatus',
-      ValueSource: 'ValueSource',
-    },
-  },
 };
 
 const modularDeprecationMessage =
@@ -530,9 +508,6 @@ export function createDeprecationProxy<T extends object>(instance: T): T {
           prop === 'FieldPath'
         ) {
           deprecationConsoleWarning('firestore', prop, 'statics', false);
-        }
-        if (prop === 'LastFetchStatus' || prop === 'ValueSource') {
-          deprecationConsoleWarning('remoteConfig', prop, 'statics', false);
         }
         if (prop === 'CustomProvider') {
           deprecationConsoleWarning('appCheck', prop, 'statics', false);
