+++ packages/app/lib/internal/constants.ts
@@ -22,7 +22,6 @@ export const DEFAULT_APP_NAME = '[DEFAULT]';
 export const KNOWN_NAMESPACES = [
   'auth',
   'analytics',
-  'remoteConfig',
   'firestore',
   'indexing',
   'naturalLanguage',
