+++ packages/remote-config/typedoc.json
@@ -1,5 +1,5 @@
 {
   "$schema": "https://typedoc.org/schema.json",
-  "entryPoints": ["lib/modular.ts", "lib/types/remote-config.ts"],
+  "entryPoints": ["lib/index.ts", "lib/types/remote-config.ts"],
   "tsconfig": "tsconfig.json"
 }
